// SISTEMA_ALFONZO_UGARTE/src/app/models/padre.model.ts
export interface Padre {
  id?: number;
  nombres: string;
  apellidos: string;
  dni: string;
  telefono?: string;
  correo?: string;
  fechaNacimiento?: string;
  estado?: boolean;
  fechaCreacion?: string;
  fechaActualizacion?: string;
  usuarioCreacion?: string;
  usuarioActualizacion?: string;
}